JSON: remove blank values from output
Yes! I really am doing this _everywhere_.

We don't want blank values in our JSON, so we're using a simple helper method to cast all blank field values to nil, which will be removed from the output by jBuilders's json.ignore_nil! option.

Gotta be a WAY classier way to do this, but for now this works.

Closes #142